Added statistic

+++ bot/data_base/requests/statistics_table.py
@@ -40,9 +40,17 @@ class StatisticsTable:
         :type picture_count: int
         :type sticker_count: int
         """
+        from bot.data_base.db import UserTable
+        print(user_id)
+        if self.is_stat(user_id) is True:
+            return
+        if not UserTable.is_user(user_id=user_id): UserTable.add_user(user_id=user_id)
         cur = self.conn.cursor()
         cur.execute(
-            f"INSERT INTO {self.table_name} ({self.user_id}, {self.message_count}, {self.picture_count}, {self.sticker_count}) VALUES ({user_id}, {message_count}, {picture_count}, {sticker_count})")
+            f"INSERT INTO {self.table_name} "
+            f"({self.user_id}, {self.message_count}, {self.picture_count}, {self.sticker_count}) "
+            f"VALUES ({user_id}, {message_count}, {picture_count}, {sticker_count})"
+        )
         self.conn.commit()
         cur.close()
 
@@ -51,6 +59,7 @@ class StatisticsTable:
         :type user_id: int
         :rtype: tuple of int
         """
+        if self.is_stat(user_id) is False: return None
         cur = self.conn.cursor()
         cur.execute(f"SELECT * FROM {self.table_name} WHERE {self.user_id} = {user_id}")
         result = cur.fetchone()
@@ -75,15 +84,21 @@ class StatisticsTable:
         :type sticker_count: int
         """
         if self.is_stat(user_id) is False:
-            add_stat(user_id, message_count, picture_count, sticker_count)
+            self.add_stat(user_id, message_count, picture_count, sticker_count)
             return
         else:
             cur = self.conn.cursor()
-            cur.execute(f"UPDATE {self.table_name} "
-                        f"SET {self.message_count} = {message_count} + {message_count}, "
-                        f"{self.picture_count} = {picture_count} + {picture_count}, "
-                        f"{self.sticker_count} = {sticker_count} + {sticker_count} "
-                        f"WHERE {self.user_id} = {user_id}")
+            stats = self.get_stat(user_id)
+            message_count += stats[2]
+            picture_count += stats[3]
+            sticker_count += stats[4]
+            cur.execute(
+                f"UPDATE {self.table_name} SET "
+                f"{self.message_count} = {message_count}, "
+                f"{self.picture_count} = {picture_count}, "
+                f"{self.sticker_count} = {sticker_count} "
+                f"WHERE {self.user_id} = {user_id}"
+            )
             self.conn.commit()
             cur.close()
 
@@ -94,17 +109,15 @@ class StatisticsTable:
         """
         cur = self.conn.cursor()
         cur.execute(f"SELECT * FROM {self.table_name} WHERE {self.user_id} = {user_id}")
-        result = cur.fetchone()
+        user = cur.fetchone()
         cur.close()
-        if result is None:
-            return False
-        else:
-            return True
+        return user is not None
 
     def delete_stat(self, user_id):
         """
         :type user_id: int
         """
+        if self.is_stat(user_id) is False: return
         cur = self.conn.cursor()
         cur.execute(f"DELETE FROM {self.table_name} WHERE {self.user_id} = {user_id}")
         self.conn.commit()
